changed to gmtl::Vec<float, 3>

+++ AlephEngine/Gravity.h
@@ -7,7 +7,7 @@ namespace AlephEngine
 	class Gravity {
 	public:
 		// m^3 / kg / s^2
-		const double G = 6.67 * pow(10, -11);
+		const float G = 6.67 * pow(10, -11);
 		Kinematics* m_myKin;
 		Gravity(Kinematics& myKin);
 		void attract(Kinematics& yourKin);
